Created abstract class moving wall.
Horizontal moving wall and vertical moving wall extend from this class.
