+++ src/main/java/nl/tudelft/model/LevelFactory.java
@@ -139,6 +139,9 @@ public class LevelFactory {
         
         walls.add(movingWall);
         walls.add(movingWall2);
+        
+        VerMovingWall movingWall3 = new VerMovingWall(resources.getSmallVWallImage(), 200, 200, -2);
+        walls.add(movingWall3);
 
         return new Level(walls, projectiles, pickups, bubbles, time, id);
     }
